Reorganize internal test logic
